[FE] jsDocs 추가
