+++ fe/src/components/Header/LanguageButton/LanguageButton.tsx
@@ -4,6 +4,13 @@ import { useTranslations } from "next-intl";
 import { useRouter, usePathname } from "next/navigation";
 import styles from "./LanguageButton.module.scss";
 
+/**
+ * 언어 전환 버튼.
+ * - 클릭 시 현재 경로의 locale 부분을 'ko' ↔ 'en' 으로 전환합니다.
+ *
+ * @param {Object} props
+ * @param {string} props.locale - 현재 언어 코드
+ */
 export default function LanguageButton({ locale }: { locale: string }) {
   const t = useTranslations("navigation");
   const router = useRouter();
