+++ fe/src/components/ScrollIndicator/ScrollIndicator.tsx
@@ -1,6 +1,16 @@
 import React from "react";
 import styles from "./ScrollIndicator.module.scss";
 
+/**
+ * 다음 섹션으로 스크롤을 안내하는 버튼.
+ * - 클릭 시 현재 화면 높이 + extraMoveHeight 만큼 스크롤.
+ *
+ * @component
+ * @param {Object} props
+ * @param {number} [props.extraMoveHeight=0] - 추가로 이동할 높이(px)
+ * @param {number} [props.bottom=80] - 버튼의 하단 위치(px)
+ * @param {boolean} [props.isDisabled=false] - 버튼 비활성화 여부
+ */
 export default function ScrollIndicator({
   extraMoveHeight = 0,
   bottom = 80,
